Specialized functions for placing text in memes

Specialized functions for placing text in memes

meme_text_distbf(img, newgirl, guy, oldgirl, ...)

meme_text_rock(img, dwayne, girl, ...)

meme_text_batman(img, robin, batman, ...)

meme_text_trump(img, rightpg, leftpg = "", ...)

Arguments

img
An image of class magick_image. See meme_get().

newgirl
A character string. Text for the left-most person in the "DistractedBf" meme.

guy
A character string. Text for the center person in the "DistractedBf" meme.

oldgirl
A character string. Text for the right-most person in the "DistractedBf" meme.

...
Arguments passed on to magick::image_annotate

color

a valid such as "navyblue" or "#000080"

gravity

string with value from gravity_types.

location

geometry string with location relative to gravity

degrees

rotates text around center point

size

font-size in pixels

font

string with font family such as "sans", "mono", "serif", "Times", "Helvetica", "Trebuchet", "Georgia", "Palatino" or "Comic Sans".

strokecolor

a adds a stroke (border around the text)

boxcolor

a for background color that annotation text is rendered on.

dwayne
A character string. Text for the top speech bubble in the "therockdriving" meme.

girl
A character string. Text for the other speech bubble in the "therockdriving" meme.

robin
A character string. Text for the Robin's speech bubble in the "batmanrobin" meme.

batman
A character string. Text for the Batman's speech bubble in the "batmanrobin" meme.

rightpg
A character string. Text for the right page in the "trumpbill" meme.

leftpg
A character string. Text for the left page in the "trumpbill" meme.

Functions

meme_text_distbf: Text function for the distracted boyfriend meme.

meme_text_rock: Text function for The Rock driving meme.

meme_text_batman: Text function for the Batman slaps Robin meme.

meme_text_trump: Text function for the Trump "first order of business" meme.

Examples

meme_get("DistractedBf") %>%
  meme_text_distbf("tidyverse", "new R users", "base R")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      500    333 sRGB       TRUE         0 28x28  

meme_get("TheRockDriving") %>%
  meme_text_rock("What's your favorite thing to do in R?" , "Write for loops.")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      351    432 sRGB       TRUE         0 28x28  

meme_get("BatmanRobin") %>%
  meme_text_batman("Hey, I'm Batman!" , "No, you idiot, I'm Batman!")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      400    387 sRGB       TRUE         0 72x72  

meme_get("TrumpBillSigning") %>%
  meme_text_trump("Stuff and nonsense")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      500    410 sRGB       TRUE         0 28x28