Posts a meme to Twitter

Pipe your meme directly into this function to post your meme on Twitter. Make sure to have rtweet installed!

meme_tweet(meme,
  tweet_text = "It's dangerous to meme alone. Here take this.",
  tag_rstatsmemes = FALSE, token = NULL)

Arguments

meme
an external pointer of class 'magick-image'

tweet_text
The text that you want to post along with your meme

tag_rstatsmemes
adds two hashtags and tags the "rstatsmemes" account

token
Every user should have their own Oauth (Twitter API) token. By default token = rtweet::get_token(), which looks for a saved Twitter token via environment variables (which is what rtweet::create_token() sets up by default during initial token creation). For instruction on how to create a Twitter token see the tokens vignette, i.e., vignettes("auth", "rtweet") or see ?tokens. See rtweet::get_token().

How to share your meme on Twitter

First

Authenticate your Twitter account, i.e. create a token)

Then

Create meme of your liking

Finally

Post meme to Twitter

References

Kearney, M. W. (2018). rtweet: Collecting Twitter Data. R package version 0.6.7 Retrieved from https://cran.r-project.org/package=rtweet

Examples

# NOT RUN {
meme_get("OprahGiveaway") %>%
  meme_text_bottom("EVERYONE GETS R MEMES!!!", size = 36) %>%
  meme_tweet(tweet_text = "It's dangerous to meme alone. Here take this.",
  tag_rstatsmemes = T)
# }