Specialized functions for placing text in memes

Specialized functions for placing text in memes

meme_text_distbf(img, newgirl, guy, oldgirl, font = "Impact",
  color = "white", strokecolor = "black", size = 44, ...)

meme_text_rock(img, dwayne, girl, font = "Impact", color = "black",
  strokecolor = NULL, size = 16, ...)

meme_text_batman(img, robin, batman, font = "Impact", color = "black",
  strokecolor = NULL, size = 20, ...)

meme_text_trump(img, rightpg, leftpg = "", font = "Impact",
  color = "black", strokecolor = NULL, size = 20, ...)

meme_text_pigeon(img, isthis, humanoid, butterfly, ...)

meme_text_chopper(img, sr1, jr1, sr2, jr2, sr3, ...)

meme_text_brain(img, br1, br2, br3, br4, ...)

meme_text_kermit(img, good, bad, ...)

meme_text_buttons(img, lbtn, rbtn, ...)

Arguments

img
An image of class magick_image. See meme_get.

newgirl
A character string. Text for the left-most person in the "DistractedBf" meme.

guy
A character string. Text for the center person in the "DistractedBf" meme.

oldgirl
A character string. Text for the right-most person in the "DistractedBf" meme.

...
Arguments passed on to magick::image_annotate

color

a valid such as "navyblue" or "#000080"

gravity

string with value from gravity_types.

location

geometry string with location relative to gravity

degrees

rotates text around center point

size

font-size in pixels

font

string with font family such as "sans", "mono", "serif", "Times", "Helvetica", "Trebuchet", "Georgia", "Palatino" or "Comic Sans".

strokecolor

a adds a stroke (border around the text)

boxcolor

a for background color that annotation text is rendered on.

dwayne
A character string. Text for the top speech bubble in the "therockdriving" meme.

girl
A character string. Text for the other speech bubble in the "therockdriving" meme.

robin
A character string. Text for the Robin's speech bubble in the "batmanrobin" meme.

batman
A character string. Text for the Batman's speech bubble in the "batmanrobin" meme.

rightpg
A character string. Text for the right page in the "trumpbill" meme.

leftpg
A character string. Text for the left page in the "trumpbill" meme.

isthis
A character string. Text for the question in the "IsThisAPigeon" meme.

humanoid
A character string. Text for the humanoid in the "IsThisAPigeon" meme.

butterfly
A character string. Text for the butterfly in the "IsThisAPigeon" meme.

sr1
A character string. Text for the first panel.

jr1
A character string. Text for the second panel.

sr2
A character string. Text for the third panel.

jr2
A character string. Text for the fourth panel.

sr3
A character string. Text for the fifth panel.

br1
A character string. Text for the first panel.

br2
A character string. Text for the second panel.

br3
A character string. Text for the third panel.

br4
A character string. Text for the fourth panel.

good
A character string. A good idea.

lbtn
A character string. The left button.

rbtn
A character string. The right button.

dark
A character string. A tempting, bad idea.

Functions

meme_text_distbf: Text function for the distracted boyfriend meme.

meme_text_rock: Text function for The Rock driving meme.

meme_text_batman: Text function for the Batman slaps Robin meme.

meme_text_trump: Text function for the Trump "first order of business" meme.

meme_text_pigeon: Text function for the Is this a pigeon? meme.

meme_text_chopper: Text function for the American Chopper Senior vs. Junior fight meme.

meme_text_brain: Text function for the expanding brain meme.

meme_text_kermit: Text function for the Off Ramp meme.

meme_text_buttons: Text function for the Two Buttons Anxiety meme.

Examples

meme_get("DistractedBf") %>%
  meme_text_distbf("tidyverse", "new R users", "base R")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      500    333 sRGB       TRUE         0 28x28  

meme_get("TheRockDriving") %>%
  meme_text_rock("What's your favorite thing to do in R?" , "Write for loops.")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      351    432 sRGB       TRUE         0 28x28  

meme_get("BatmanRobin") %>%
  meme_text_batman("Hey, I'm Batman!" , "No, you idiot, I'm Batman!")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      400    387 sRGB       TRUE         0 72x72  

meme_get("TrumpBillSigning") %>%
  meme_text_trump("Stuff and nonsense")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      500    410 sRGB       TRUE         0 28x28  

meme_get("IsThisAPigeon") %>%
  meme_text_pigeon("Is this a pigeon?" , "Me", "Not a pigeon")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      700    395 sRGB       TRUE         0 72x72  

meme_get("AmericanChopper") %>%
  meme_text_chopper("Stop calling yourself 'just an R user'!",
                    "But I've never written a package!",
                    "So?? You're still an important part of the community!",
                    "But people who write packages are more important",
                    "NO! All members of the community are valued & deserve respect!")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      356   1000 sRGB       TRUE         0 28x28  

meme_get("ExpandingBrain") %>%
  meme_text_brain("text1",
                    "text2",
                    "text3",
                    "text4")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      500    701 sRGB       TRUE         0 72x72  

meme_get("EvilKermit") %>%
  meme_text_kermit("Me: Your text here", "Me: Your evil text here")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      600    516 sRGB       TRUE         0 72x72  

meme_get("TwoButtonsAnxiety") %>%
  meme_text_buttons("Should I do this?", "Or this?")

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG      333    504 sRGB       TRUE         0 28x28